add buffer size in console

+++ cipher.py
@@ -3,7 +3,14 @@ import pyAesCrypt
 
 
 class Crypto:
-    def file_crypt(self, file: str, password: str, buffer_size: int = 524288, logger: bool = True) -> None:
+    def file_crypt(
+        self,
+        file: str,
+        password: str,
+        buffer_size: int = 524288,
+        logger: bool = True,
+        original_delete: bool = True,
+    ) -> None:
         """
         Encrypts a single file using pyAesCrypt.
 
@@ -12,6 +19,7 @@ class Crypto:
             password (str): The password to use for encryption.
             buffer_size (int, optional): The size of the encryption buffer.
             logger (bool, optional): If True, print the path of the encrypted file.
+            original_delete (bool, optional): Deleting the original file.
 
         Returns:
             None
@@ -19,9 +27,17 @@ class Crypto:
         pyAesCrypt.encryptFile(str(file), str(file) + ".crp", password, buffer_size)
         if logger:
             print(f"[Encrypt] '{str(file)}.crp'")
-        os.remove(file)
+        if original_delete:
+            os.remove(file)
 
-    def file_decrypt(self, file: str, password: str, buffer_size: int = 524288, logger: bool = True) -> None:
+    def file_decrypt(
+        self,
+        file: str,
+        password: str,
+        buffer_size: int = 524288,
+        logger: bool = True,
+        original_delete: bool = True,
+    ) -> None:
         """
         Decrypts a single file using pyAesCrypt.
 
@@ -30,6 +46,7 @@ class Crypto:
             password (str): The password to use for decryption.
             buffer_size (int, optional): The size of the decryption buffer.
             logger (bool, optional): If True, print the path of the decrypted file.
+            original_delete (bool, optional): Deleting the original file.
 
         Returns:
             None
@@ -39,10 +56,17 @@ class Crypto:
         )
         if logger:
             print(f"[Decrypt] '{str(os.path.splitext(file)[0])}'")
-        os.remove(file)
+        if original_delete:
+            os.remove(file)
 
     def directory_crypt(
-        self, directory: str, password: str, buffer_size: int = 524288, logger: bool = True, title_logger: bool = True
+        self,
+        directory: str,
+        password: str,
+        buffer_size: int = 524288,
+        logger: bool = True,
+        title_logger: bool = True,
+        original_delete: bool = True,
     ) -> None:
         """
         Recursively encrypts all files in a directory using pyAesCrypt.
@@ -53,23 +77,32 @@ class Crypto:
             buffer_size (int, optional): The size of the encryption buffer.
             logger (bool, optional): If True, print the path of each file as it is encrypted.
             title_logger (bool, optional): If True, print a title before starting the encryption process.
+            original_delete (bool, optional): Deleting the original file.
 
         Returns:
             None
         """
         if logger and title_logger:
-            print("[***Crypto***]")
+            print("[--Crypto--]")
         if logger:
             print(f"[Directory] '{directory}'")
         for name in os.listdir(directory):
             path = os.path.join(directory, name)
             if os.path.isfile(path):
-                self.file_crypt(path, password, buffer_size, logger)
+                self.file_crypt(path, password, buffer_size, logger, original_delete)
             else:
-                self.directory_crypt(path, password, buffer_size, logger, False)
+                self.directory_crypt(
+                    path, password, buffer_size, logger, False, original_delete
+                )
 
     def directory_decrypt(
-        self, directory: str, password: str, buffer_size: int = 524288, logger: bool = True, title_logger: bool = True
+        self,
+        directory: str,
+        password: str,
+        buffer_size: int = 524288,
+        logger: bool = True,
+        title_logger: bool = True,
+        original_delete: bool = True,
     ) -> None:
         """
         Recursively decrypts all files in a directory using pyAesCrypt.
@@ -80,20 +113,25 @@ class Crypto:
             buffer_size (int, optional): The size of the decryption buffer.
             logger (bool, optional): If True, print the path of each file as it is decrypted.
             title_logger (bool, optional): If True, print a title before starting the decryption process.
+            original_delete (bool, optional): Deleting the original file.
 
         Returns:
             None
         """
         if logger and title_logger:
-            print("[***Derypto***]")
+            print("[--Derypto--]")
         if logger:
             print(f"[Directory] '{directory}'")
         for name in os.listdir(directory):
             path = os.path.join(directory, name)
             if os.path.isfile(path):
                 try:
-                    self.file_decrypt(path, password, buffer_size, logger)
+                    self.file_decrypt(
+                        path, password, buffer_size, logger, original_delete
+                    )
                 except Exception:
                     pass
             else:
-                self.directory_decrypt(path, password, buffer_size, logger, False)
+                self.directory_decrypt(
+                    path, password, buffer_size, logger, False, original_delete
+                )
